pyproject.toml
[tool.poetry]
name = "talecraft"
version = "0.1.0"
description = ""
authors = ["bakatrouble <bakatrouble@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
eth-brownie = "^1.16.4"
Wand = "^0.6.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
